backend/requirements.txt
# Same as requirements.txt but with Windows-compatible versions
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
tensorflow-cpu==2.13.0  # Use CPU version for easier Windows setup
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.0
scikit-learn==1.3.0
mediapipe==0.10.0
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0